fasthtml_openai/requirements.txt
openai==1.51.0
libsql-client==0.3.1
libsql-experimental==0.0.40
python-dotenv==1.0.1
python-fasthtml==0.6.9